requirements.txt
torch~=1.10
torchvision~=0.11.1
pandas~=1.3.4
pillow~=8.4.0
numpy~=1.21.2
pycocotools~=2.0.3
matplotlib~=3.5.0
tensorboard~=2.7